+++ src/main/java/ca/yorku/cse2311/tab2pdf/model/ILongDraw.java
@@ -12,5 +12,5 @@ import com.itextpdf.text.pdf.PdfWriter;
  */
 public interface ILongDraw extends IDrawable {
 
-    public void drawLong(int staveNumber, int lineNumber, int xCoordinate, PdfWriter writer, int oldStave, int OldLine, int oldXCoordinate, String oldString);
+    public void drawLong(int staveNumber, int lineNumber, float xCoordinate, PdfWriter writer, int oldStave, int OldLine, float oldXCoordinate, String oldString);
 }
